requirements.txt
# ---------------------------
# Core
# ---------------------------
python-dotenv==1.0.1

# ---------------------------
# Streamlit (Web UI)
# ---------------------------
streamlit==1.32.2

# ---------------------------
# LangChain (Stable 0.1.x)
# ---------------------------
langchain==0.1.20
langchain-core==0.1.53
langchain-community==0.0.38
langchain-text-splitters==0.0.2
langchain-huggingface==0.0.3

# ---------------------------
# Vector Store
# ---------------------------
faiss-cpu==1.7.4

# ---------------------------
# Embeddings
# ---------------------------
sentence-transformers==2.2.2
torch==2.1.2

# ---------------------------
# PDF Loading
# ---------------------------
pypdf==3.17.4

# ---------------------------
# Utilities
# ---------------------------
numpy==1.26.4
tqdm==4.66.1

